Move ea_custom_dag to dags subfolder.
