+++ ea_airflow_util/__init__.py
@@ -1,4 +1,4 @@
-from ea_airflow_util.dags.dag_util.ea_custom_dag import EACustomDAG
+from ea_airflow_util.dags.ea_custom_dag import EACustomDAG
 
 from ea_airflow_util.dags.aws_param_store_to_airflow_dag import AWSParamStoreToAirflowDAG
 from ea_airflow_util.dags.run_dbt_airflow_dag import RunDbtDag
